pyproject.toml
[tool.poetry]
name = "cryptodivebot"
version = "0.1.0"
description = "Telegram Mini App for dating"
authors = ["Your Name"]

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.13.1"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.0"